Made gridworld a prefab so we can have more of them at once
