+++ UnitySDK/Assets/ML-Agents/Examples/GridWorld/Prefabs/Area.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5c2bd19e4bbda4991b74387ca5d28156
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
