--- UnitySDK/Assets/ML-Agents/Examples/GridWorld/Prefabs/agent.prefab.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 628960e910f094ad1909ecc88cc8016d
-timeCreated: 1504978468
-licenseType: Pro
-NativeFormatImporter:
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
