+++ UnitySDK/Assets/ML-Agents/Examples/GridWorld/Scripts/GridAgent.cs
@@ -5,8 +5,9 @@ using MLAgents;
 
 public class GridAgent : Agent
 {
+    private Academy m_Academy;
     [Header("Specific to GridWorld")]
-    private GridAcademy m_Academy;
+    public GridArea m_Area;
     public float timeBetweenDecisionsAtInference;
     private float m_TimeSinceDecision;
 
@@ -27,7 +28,7 @@ public class GridAgent : Agent
 
     public override void InitializeAgent()
     {
-        m_Academy = FindObjectOfType(typeof(GridAcademy)) as GridAcademy;
+        m_Academy = FindObjectOfType<Academy>();
     }
 
     public override void CollectObservations()
@@ -50,7 +51,7 @@ public class GridAgent : Agent
         // Prevents the agent from picking an action that would make it collide with a wall
         var positionX = (int)transform.position.x;
         var positionZ = (int)transform.position.z;
-        var maxPosition = m_Academy.gridSize - 1;
+        var maxPosition = (int)m_Academy.resetParameters["gridSize"] - 1;
 
         if (positionX == 0)
         {
@@ -123,7 +124,7 @@ public class GridAgent : Agent
     // to be implemented by the developer
     public override void AgentReset()
     {
-        m_Academy.AcademyReset();
+        m_Area.AreaReset();
     }
 
     public void FixedUpdate()
