+++ UnitySDK/Assets/ML-Agents/Examples/GridWorld/Scripts/GridArea.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 676658555cb2d4884aa8285062aab2a1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
